+++ Realization/frontend/czechidm-core/src/components/advanced/Form/ShortTextFormAttributeRenderer.js
@@ -1,0 +1,109 @@
+import _ from 'lodash';
+import Joi from 'joi';
+//
+import TextFormAttributeRenderer from './TextFormAttributeRenderer';
+
+/**
+ * Short text form value component
+ * - supports multiple and confidential attributes
+ *
+ * @author Radek Tomiška
+ */
+export default class ShortTextFormAttributeRenderer extends TextFormAttributeRenderer {
+
+  /**
+   * Returns true, when confidential mode is supported
+   *
+   * @return {boolean}
+   */
+  supportsConfidential() {
+    return true;
+  }
+
+  /**
+   * Returns true, when multi value mode is supported
+   *
+   * @return {boolean}
+   */
+  supportsMultiple() {
+    return true;
+  }
+
+  /**
+   * Fill form value field by persistent type from input value
+   *
+   * @param  {FormValue} formValue - form value
+   * @param  {[type]} formComponent
+   * @return {FormValue}
+   */
+  fillFormValue(formValue, rawValue) {
+    formValue.shortTextValue = rawValue;
+    return formValue;
+  }
+
+  /**
+   * Returns value to ipnut from given (persisted) form value
+   *
+   * @param  {FormValue} formValue
+   * @return {object} value by persistent type
+   */
+  getInputValue(formValue) {
+    return formValue.shortTextValue ? formValue.shortTextValue : formValue.value;
+  }
+
+  /**
+   * Return value by attribute persistent type
+   *
+   * @param  {FormValue} formValue form value
+   * @return {oneOf([string, boolean, long])}
+   */
+  toInputValues(formValues) {
+    const { attribute } = this.props;
+    //
+    if (formValues === null) {
+      return attribute.defaultValue;
+    }
+    if (_.isArray(formValues)) {
+      // multi values are transformed to multi lines
+      let result = null;
+      formValues.forEach(singleValue => {
+        if (result !== null) {
+          if (!attribute.multiple) {
+            return false;
+          }
+          result += '\n';
+        }
+        const inputValue = this.toInputValues(singleValue);
+        if (inputValue) {
+          if (result === null) { // single values should not be concated
+            result = inputValue;
+          } else {
+            result += inputValue;
+          }
+        }
+      });
+      return result;
+    }
+    //
+    if (attribute.confidential) {
+      return formValues.shortTextValue; // returns proxied guarded string everytime
+    }
+    return this.getInputValue(formValues);
+  }
+
+  /**
+   * Returns joi validator by persistent type
+   *
+   * @param  {FormAttribute} attribute
+   * @return {Joi}
+   */
+  getInputValidation() {
+    const { attribute } = this.props;
+    //
+    let validation = Joi.string().max(2000);
+    if (!attribute.required) {
+      validation = validation.concat(Joi.string().allow(null).allow(''));
+    }
+    return validation;
+  }
+}